refactor: split cli.rs into cloud.rs and enterprise.rs modules
- Extract Cloud enums to cli/cloud.rs (1477 lines)
- Extract Enterprise enums to cli/enterprise.rs (1389 lines)
- Keep shared types in cli/mod.rs (404 lines)
- Improves maintainability and navigation

Closes #440
