--- crates/redisctl/src/cli.rs
@@ -1,3261 +1,0 @@
-//! CLI structure and command definitions
-//!
-//! Defines the command-line interface using clap with a three-layer architecture:
-//! 1. Raw API access (`api` commands)
-//! 2. Human-friendly interface (`cloud`/`enterprise` commands)
-//! 3. Workflow orchestration (`workflow` commands - future)
-
-use clap::{Parser, Subcommand};
-use redisctl_config::DeploymentType;
-
-/// Redis management CLI with unified access to Cloud and Enterprise
-#[derive(Parser, Debug)]
-#[command(name = "redisctl")]
-#[command(
-    version,
-    about = "Redis management CLI for Cloud and Enterprise deployments"
-)]
-#[command(long_about = "
-Redis management CLI for Cloud and Enterprise deployments
-
-EXAMPLES:
-    # Set up a Cloud profile
-    redisctl profile set mycloud --type cloud --api-key KEY --api-secret SECRET
-
-    # Set up an Enterprise profile
-    redisctl profile set myenterprise --type enterprise --url https://cluster:9443 --username admin
-
-    # List databases
-    redisctl cloud database list
-    redisctl enterprise database list
-
-    # Get JSON output for scripting
-    redisctl cloud subscription list -o json
-
-    # Filter output with JMESPath
-    redisctl cloud database list -q 'databases[?status==`active`]'
-
-    # Direct API access
-    redisctl api cloud get /subscriptions
-    redisctl api enterprise get /v1/cluster
-
-For more help on a specific command, run:
-    redisctl <command> --help
-")]
-pub struct Cli {
-    /// Profile to use for this command
-    #[arg(long, short, global = true, env = "REDISCTL_PROFILE")]
-    pub profile: Option<String>,
-
-    /// Path to alternate configuration file
-    #[arg(long, global = true, env = "REDISCTL_CONFIG_FILE")]
-    pub config_file: Option<String>,
-
-    /// Output format
-    #[arg(long, short = 'o', global = true, value_enum, default_value = "auto")]
-    pub output: OutputFormat,
-
-    /// JMESPath query to filter output
-    #[arg(long, short = 'q', global = true)]
-    pub query: Option<String>,
-
-    /// Enable verbose logging
-    #[arg(long, short, global = true, action = clap::ArgAction::Count)]
-    pub verbose: u8,
-
-    #[command(subcommand)]
-    pub command: Commands,
-}
-
-/// Output format options
-#[derive(Debug, Clone, Copy, clap::ValueEnum)]
-pub enum OutputFormat {
-    /// Automatically choose format based on command and context
-    Auto,
-    /// JSON output
-    Json,
-    /// YAML output
-    Yaml,
-    /// Human-readable table format
-    Table,
-}
-
-/// Top-level commands
-#[derive(Subcommand, Debug)]
-pub enum Commands {
-    /// Raw API access - direct REST endpoint calls
-    #[command(name = "api")]
-    #[command(after_help = "EXAMPLES:
-    # GET request to Cloud API
-    redisctl api cloud get /subscriptions
-
-    # GET request to Enterprise API
-    redisctl api enterprise get /v1/cluster
-
-    # POST request with JSON data
-    redisctl api cloud post /subscriptions --data '{\"name\":\"my-sub\"}'
-
-    # POST request from file
-    redisctl api cloud post /subscriptions --data @subscription.json
-
-    # Output as JSON for scripting
-    redisctl api enterprise get /v1/bdbs -o json
-")]
-    Api {
-        /// Platform type (cloud or enterprise)
-        #[arg(value_enum)]
-        deployment: DeploymentType,
-
-        /// HTTP method
-        #[arg(value_parser = parse_http_method)]
-        method: HttpMethod,
-
-        /// API endpoint path (e.g., /subscriptions)
-        path: String,
-
-        /// Request body (JSON string or @file)
-        #[arg(long)]
-        data: Option<String>,
-    },
-
-    /// Profile management
-    #[command(subcommand, visible_alias = "prof", visible_alias = "pr")]
-    #[command(after_help = "EXAMPLES:
-    # Create a Cloud profile
-    redisctl profile set mycloud --type cloud --api-key KEY --api-secret SECRET
-
-    # Create an Enterprise profile
-    redisctl profile set myenterprise --type enterprise --url https://cluster:9443 --username admin
-
-    # List all profiles
-    redisctl profile list
-
-    # Show profile details
-    redisctl profile show mycloud
-
-    # Validate configuration
-    redisctl profile validate
-
-    # Set default profiles
-    redisctl profile default-cloud mycloud
-    redisctl profile default-enterprise myenterprise
-")]
-    Profile(ProfileCommands),
-
-    /// Cloud-specific operations
-    #[command(subcommand, visible_alias = "cl")]
-    Cloud(CloudCommands),
-
-    /// Enterprise-specific operations
-    #[command(subcommand, visible_alias = "ent", visible_alias = "en")]
-    Enterprise(EnterpriseCommands),
-
-    /// Files.com API key management (for support package uploads)
-    #[command(subcommand, visible_alias = "fk")]
-    FilesKey(FilesKeyCommands),
-
-    /// Version information
-    #[command(visible_alias = "ver", visible_alias = "v")]
-    Version,
-
-    /// Generate shell completions
-    #[command(visible_alias = "comp")]
-    Completions {
-        /// Shell to generate completions for
-        #[arg(value_enum)]
-        shell: Shell,
-    },
-}
-
-/// Supported shells for completion generation
-#[derive(Debug, Clone, Copy, clap::ValueEnum)]
-#[allow(clippy::enum_variant_names)]
-pub enum Shell {
-    /// Bourne Again Shell
-    Bash,
-    /// Z Shell
-    Zsh,
-    /// Friendly Interactive Shell
-    Fish,
-    /// PowerShell
-    #[value(name = "powershell", alias = "power-shell")]
-    PowerShell,
-    /// Elvish
-    Elvish,
-}
-
-/// HTTP methods for raw API access
-#[derive(Debug, Clone)]
-pub enum HttpMethod {
-    Get,
-    Post,
-    Put,
-    Patch,
-    Delete,
-}
-
-/// Parse HTTP method case-insensitively
-fn parse_http_method(s: &str) -> Result<HttpMethod, String> {
-    match s.to_lowercase().as_str() {
-        "get" => Ok(HttpMethod::Get),
-        "post" => Ok(HttpMethod::Post),
-        "put" => Ok(HttpMethod::Put),
-        "patch" => Ok(HttpMethod::Patch),
-        "delete" => Ok(HttpMethod::Delete),
-        _ => Err(format!(
-            "invalid HTTP method: {} (valid: get, post, put, patch, delete)",
-            s
-        )),
-    }
-}
-
-impl std::fmt::Display for HttpMethod {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            HttpMethod::Get => write!(f, "GET"),
-            HttpMethod::Post => write!(f, "POST"),
-            HttpMethod::Put => write!(f, "PUT"),
-            HttpMethod::Patch => write!(f, "PATCH"),
-            HttpMethod::Delete => write!(f, "DELETE"),
-        }
-    }
-}
-
-/// Profile management commands
-#[derive(Subcommand, Debug)]
-pub enum ProfileCommands {
-    /// List all configured profiles
-    #[command(visible_alias = "ls", visible_alias = "l")]
-    List,
-
-    /// Show the path to the configuration file
-    Path,
-
-    /// Show details of a specific profile
-    #[command(visible_alias = "sh", visible_alias = "get")]
-    Show {
-        /// Profile name to show
-        name: String,
-    },
-
-    /// Set or create a profile
-    #[command(visible_alias = "add", visible_alias = "create")]
-    #[command(after_help = "EXAMPLES:
-    # Create a Cloud profile
-    redisctl profile set mycloud --type cloud \\
-        --api-key A3qcymrvqpn9rrgdt40sv5f9yfxob26vx64hwddh8vminqnkgfq \\
-        --api-secret S3s8ecrrnaguqkvwfvealoe3sn25zqs4wc4lwgo4rb0ud3qm77c
-
-    # Create an Enterprise profile (password will be prompted)
-    redisctl profile set prod --type enterprise \\
-        --url https://cluster.example.com:9443 \\
-        --username admin@example.com
-
-    # Create Enterprise profile with password
-    redisctl profile set staging --type enterprise \\
-        --url https://staging:9443 \\
-        --username admin \\
-        --password mypassword
-
-    # Create Enterprise profile allowing insecure connections
-    redisctl profile set local --type enterprise \\
-        --url https://localhost:9443 \\
-        --username admin@redis.local \\
-        --insecure
-")]
-    Set {
-        /// Profile name
-        name: String,
-
-        /// Platform type: 'cloud' for Redis Cloud or 'enterprise' for Redis Enterprise
-        #[arg(long, value_enum, visible_alias = "deployment")]
-        r#type: DeploymentType,
-
-        /// API key (for Cloud profiles)
-        #[arg(long, required_if_eq("type", "cloud"))]
-        api_key: Option<String>,
-
-        /// API secret (for Cloud profiles)
-        #[arg(long, required_if_eq("type", "cloud"))]
-        api_secret: Option<String>,
-
-        /// API URL (for Cloud profiles)
-        #[arg(long, default_value = "https://api.redislabs.com/v1")]
-        api_url: String,
-
-        /// Enterprise URL (for Enterprise profiles)
-        #[arg(long, required_if_eq("type", "enterprise"))]
-        url: Option<String>,
-
-        /// Username (for Enterprise profiles)
-        #[arg(long, required_if_eq("type", "enterprise"))]
-        username: Option<String>,
-
-        /// Password (for Enterprise profiles)
-        #[arg(long)]
-        password: Option<String>,
-
-        /// Allow insecure connections (for Enterprise profiles)
-        #[arg(long)]
-        insecure: bool,
-
-        /// Store credentials in OS keyring instead of config file
-        #[cfg(feature = "secure-storage")]
-        #[arg(long)]
-        use_keyring: bool,
-    },
-
-    /// Remove a profile
-    #[command(visible_alias = "rm", visible_alias = "del", visible_alias = "delete")]
-    Remove {
-        /// Profile name to remove
-        name: String,
-    },
-
-    /// Set the default profile for enterprise commands
-    #[command(name = "default-enterprise", visible_alias = "def-ent")]
-    DefaultEnterprise {
-        /// Profile name to set as default for enterprise commands
-        name: String,
-    },
-
-    /// Set the default profile for cloud commands
-    #[command(name = "default-cloud", visible_alias = "def-cloud")]
-    DefaultCloud {
-        /// Profile name to set as default for cloud commands
-        name: String,
-    },
-
-    /// Validate configuration file and profiles
-    #[command(visible_alias = "check")]
-    #[command(after_help = "EXAMPLES:
-    # Validate all profiles and configuration
-    redisctl profile validate
-
-    # Example output:
-    # Configuration file: /Users/user/.config/redisctl/config.toml
-    # ✓ Configuration file exists and is readable
-    # ✓ Found 2 profile(s)
-    #
-    # Profile 'mycloud' (cloud): ✓ Valid
-    # Profile 'myenterprise' (enterprise): ✓ Valid
-    #
-    # ✓ Default enterprise profile: myenterprise
-    # ✓ Default cloud profile: mycloud
-    #
-    # ✓ Configuration is valid
-")]
-    Validate,
-}
-
-/// Files.com API key management commands
-#[derive(Subcommand, Debug)]
-pub enum FilesKeyCommands {
-    /// Store Files.com API key (globally or in config)
-    #[command(visible_alias = "add")]
-    Set {
-        /// The Files.com API key provided by Redis Support
-        api_key: String,
-
-        /// Store in system keyring (most secure - recommended)
-        #[cfg(feature = "secure-storage")]
-        #[arg(long)]
-        use_keyring: bool,
-
-        /// Store globally in config file (plaintext - not recommended)
-        #[arg(long, conflicts_with = "use_keyring")]
-        global: bool,
-
-        /// Store in specific profile's config (plaintext - not recommended)
-        #[arg(long, conflicts_with_all = ["use_keyring", "global"])]
-        profile: Option<String>,
-    },
-
-    /// Get the currently configured Files.com API key
-    #[command(visible_alias = "show")]
-    Get {
-        /// Show for specific profile
-        #[arg(long)]
-        profile: Option<String>,
-    },
-
-    /// Remove Files.com API key
-    #[command(visible_alias = "rm", visible_alias = "delete")]
-    Remove {
-        /// Remove from keyring
-        #[cfg(feature = "secure-storage")]
-        #[arg(long)]
-        keyring: bool,
-
-        /// Remove from global config
-        #[arg(long, conflicts_with = "keyring")]
-        global: bool,
-
-        /// Remove from specific profile
-        #[arg(long, conflicts_with_all = ["keyring", "global"])]
-        profile: Option<String>,
-    },
-}
-
-/// Cloud Connectivity Commands
-#[derive(Subcommand, Debug)]
-pub enum CloudConnectivityCommands {
-    /// VPC Peering operations
-    #[command(subcommand, name = "vpc-peering")]
-    VpcPeering(VpcPeeringCommands),
-    /// Private Service Connect operations
-    #[command(subcommand, name = "psc")]
-    Psc(PscCommands),
-    /// Transit Gateway operations
-    #[command(subcommand, name = "tgw")]
-    Tgw(TgwCommands),
-    /// AWS PrivateLink operations
-    #[command(subcommand, name = "privatelink")]
-    PrivateLink(PrivateLinkCommands),
-}
-
-/// VPC Peering Commands
-#[derive(Subcommand, Debug)]
-pub enum VpcPeeringCommands {
-    /// Get VPC peering details
-    Get {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-    },
-    /// Create VPC peering
-    Create {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-        /// Configuration JSON file or string (use @filename for file)
-        data: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Update VPC peering
-    Update {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-        /// Peering ID
-        #[arg(long)]
-        peering_id: i32,
-        /// Configuration JSON file or string (use @filename for file)
-        data: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Delete VPC peering
-    Delete {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-        /// Peering ID
-        #[arg(long)]
-        peering_id: i32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// List Active-Active VPC peerings
-    #[command(name = "list-aa")]
-    ListActiveActive {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-    },
-    /// Create Active-Active VPC peering
-    #[command(name = "create-aa")]
-    CreateActiveActive {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-        /// Configuration JSON file or string (use @filename for file)
-        data: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Update Active-Active VPC peering
-    #[command(name = "update-aa")]
-    UpdateActiveActive {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-        /// Peering ID
-        #[arg(long)]
-        peering_id: i32,
-        /// Configuration JSON file or string (use @filename for file)
-        data: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Delete Active-Active VPC peering
-    #[command(name = "delete-aa")]
-    DeleteActiveActive {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-        /// Peering ID
-        #[arg(long)]
-        peering_id: i32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-}
-
-/// Private Service Connect (PSC) Commands
-#[derive(Subcommand, Debug)]
-pub enum PscCommands {
-    // Standard PSC Service operations
-    /// Get PSC service details
-    #[command(name = "service-get")]
-    ServiceGet {
-        /// Subscription ID
-        subscription_id: i32,
-    },
-    /// Create PSC service
-    #[command(name = "service-create")]
-    ServiceCreate {
-        /// Subscription ID
-        subscription_id: i32,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Delete PSC service
-    #[command(name = "service-delete")]
-    ServiceDelete {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Skip confirmation prompt
-        #[arg(short, long)]
-        yes: bool,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    // Standard PSC Endpoint operations
-    /// List PSC endpoints
-    #[command(name = "endpoints-list")]
-    EndpointsList {
-        /// Subscription ID
-        subscription_id: i32,
-    },
-    /// Create PSC endpoint
-    #[command(name = "endpoint-create")]
-    EndpointCreate {
-        /// Subscription ID
-        subscription_id: i32,
-        /// JSON file with endpoint configuration (use @filename or - for stdin)
-        file: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Update PSC endpoint
-    #[command(name = "endpoint-update")]
-    EndpointUpdate {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Endpoint ID
-        endpoint_id: i32,
-        /// JSON file with endpoint configuration (use @filename or - for stdin)
-        file: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Delete PSC endpoint
-    #[command(name = "endpoint-delete")]
-    EndpointDelete {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Endpoint ID
-        endpoint_id: i32,
-        /// Skip confirmation prompt
-        #[arg(short, long)]
-        yes: bool,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Get PSC endpoint creation script
-    #[command(name = "endpoint-creation-script")]
-    EndpointCreationScript {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Endpoint ID
-        endpoint_id: i32,
-    },
-    /// Get PSC endpoint deletion script
-    #[command(name = "endpoint-deletion-script")]
-    EndpointDeletionScript {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Endpoint ID
-        endpoint_id: i32,
-    },
-
-    // Active-Active PSC Service operations
-    /// Get Active-Active PSC service details
-    #[command(name = "aa-service-get")]
-    AaServiceGet {
-        /// Subscription ID
-        subscription_id: i32,
-    },
-    /// Create Active-Active PSC service
-    #[command(name = "aa-service-create")]
-    AaServiceCreate {
-        /// Subscription ID
-        subscription_id: i32,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Delete Active-Active PSC service
-    #[command(name = "aa-service-delete")]
-    AaServiceDelete {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Skip confirmation prompt
-        #[arg(short, long)]
-        yes: bool,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    // Active-Active PSC Endpoint operations
-    /// List Active-Active PSC endpoints
-    #[command(name = "aa-endpoints-list")]
-    AaEndpointsList {
-        /// Subscription ID
-        subscription_id: i32,
-    },
-    /// Create Active-Active PSC endpoint
-    #[command(name = "aa-endpoint-create")]
-    AaEndpointCreate {
-        /// Subscription ID
-        subscription_id: i32,
-        /// JSON file with endpoint configuration (use @filename or - for stdin)
-        file: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Delete Active-Active PSC endpoint
-    #[command(name = "aa-endpoint-delete")]
-    AaEndpointDelete {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Region ID
-        region_id: i32,
-        /// Endpoint ID
-        endpoint_id: i32,
-        /// Skip confirmation prompt
-        #[arg(short, long)]
-        yes: bool,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-}
-
-/// Transit Gateway (TGW) Commands
-#[derive(Subcommand, Debug)]
-pub enum TgwCommands {
-    // Standard TGW Attachment operations
-    /// List TGW attachments
-    #[command(name = "attachments-list")]
-    AttachmentsList {
-        /// Subscription ID
-        subscription_id: i32,
-    },
-    /// Create TGW attachment
-    #[command(name = "attachment-create")]
-    AttachmentCreate {
-        /// Subscription ID
-        subscription_id: i32,
-        /// JSON file with attachment configuration (use @filename or - for stdin)
-        file: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Create TGW attachment with ID in path
-    #[command(name = "attachment-create-with-id")]
-    AttachmentCreateWithId {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Transit Gateway ID
-        tgw_id: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Update TGW attachment CIDRs
-    #[command(name = "attachment-update")]
-    AttachmentUpdate {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Attachment ID
-        attachment_id: String,
-        /// JSON file with CIDR configuration (use @filename or - for stdin)
-        file: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Delete TGW attachment
-    #[command(name = "attachment-delete")]
-    AttachmentDelete {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Attachment ID
-        attachment_id: String,
-        /// Skip confirmation prompt
-        #[arg(short, long)]
-        yes: bool,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    // Standard TGW Invitation operations
-    /// List TGW resource share invitations
-    #[command(name = "invitations-list")]
-    InvitationsList {
-        /// Subscription ID
-        subscription_id: i32,
-    },
-    /// Accept TGW resource share invitation
-    #[command(name = "invitation-accept")]
-    InvitationAccept {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Invitation ID
-        invitation_id: String,
-    },
-    /// Reject TGW resource share invitation
-    #[command(name = "invitation-reject")]
-    InvitationReject {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Invitation ID
-        invitation_id: String,
-    },
-
-    // Active-Active TGW Attachment operations
-    /// List Active-Active TGW attachments
-    #[command(name = "aa-attachments-list")]
-    AaAttachmentsList {
-        /// Subscription ID
-        subscription_id: i32,
-    },
-    /// Create Active-Active TGW attachment
-    #[command(name = "aa-attachment-create")]
-    AaAttachmentCreate {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Region ID
-        region_id: i32,
-        /// JSON file with attachment configuration (use @filename or - for stdin)
-        file: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Update Active-Active TGW attachment CIDRs
-    #[command(name = "aa-attachment-update")]
-    AaAttachmentUpdate {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Region ID
-        region_id: i32,
-        /// Attachment ID
-        attachment_id: String,
-        /// JSON file with CIDR configuration (use @filename or - for stdin)
-        file: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Delete Active-Active TGW attachment
-    #[command(name = "aa-attachment-delete")]
-    AaAttachmentDelete {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Region ID
-        region_id: i32,
-        /// Attachment ID
-        attachment_id: String,
-        /// Skip confirmation prompt
-        #[arg(short, long)]
-        yes: bool,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    // Active-Active TGW Invitation operations
-    /// List Active-Active TGW resource share invitations
-    #[command(name = "aa-invitations-list")]
-    AaInvitationsList {
-        /// Subscription ID
-        subscription_id: i32,
-    },
-    /// Accept Active-Active TGW resource share invitation
-    #[command(name = "aa-invitation-accept")]
-    AaInvitationAccept {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Region ID
-        region_id: i32,
-        /// Invitation ID
-        invitation_id: String,
-    },
-    /// Reject Active-Active TGW resource share invitation
-    #[command(name = "aa-invitation-reject")]
-    AaInvitationReject {
-        /// Subscription ID
-        subscription_id: i32,
-        /// Region ID
-        region_id: i32,
-        /// Invitation ID
-        invitation_id: String,
-    },
-}
-
-/// AWS PrivateLink Commands
-#[derive(Subcommand, Debug)]
-pub enum PrivateLinkCommands {
-    /// Get PrivateLink configuration
-    Get {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-        /// Region ID (for Active-Active databases)
-        #[arg(long)]
-        region: Option<i32>,
-    },
-    /// Create PrivateLink
-    Create {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-        /// Region ID (for Active-Active databases)
-        #[arg(long)]
-        region: Option<i32>,
-        /// Configuration JSON file or string (use @filename for file)
-        data: String,
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Add principals to PrivateLink
-    #[command(name = "add-principal")]
-    AddPrincipal {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-        /// Region ID (for Active-Active databases)
-        #[arg(long)]
-        region: Option<i32>,
-        /// Configuration JSON file or string (use @filename for file)
-        data: String,
-    },
-    /// Remove principals from PrivateLink
-    #[command(name = "remove-principal")]
-    RemovePrincipal {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-        /// Region ID (for Active-Active databases)
-        #[arg(long)]
-        region: Option<i32>,
-        /// Configuration JSON file or string (use @filename for file)
-        data: String,
-    },
-    /// Get VPC endpoint creation script
-    #[command(name = "get-script")]
-    GetScript {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-        /// Region ID (for Active-Active databases)
-        #[arg(long)]
-        region: Option<i32>,
-    },
-}
-
-/// Cloud Task Commands
-#[derive(Subcommand, Debug)]
-pub enum CloudTaskCommands {
-    /// Get task status and details
-    Get {
-        /// Task ID (UUID format)
-        id: String,
-    },
-    /// Wait for task to complete
-    Wait {
-        /// Task ID (UUID format)
-        id: String,
-        /// Maximum time to wait in seconds
-        #[arg(long, default_value = "300")]
-        timeout: u64,
-        /// Polling interval in seconds
-        #[arg(long, default_value = "2")]
-        interval: u64,
-    },
-    /// Poll task status with live updates
-    Poll {
-        /// Task ID (UUID format)
-        id: String,
-        /// Polling interval in seconds
-        #[arg(long, default_value = "2")]
-        interval: u64,
-        /// Maximum number of polls (0 = unlimited)
-        #[arg(long, default_value = "0")]
-        max_polls: u64,
-    },
-}
-
-/// Cloud Fixed Database Commands
-#[derive(Subcommand, Debug)]
-pub enum CloudFixedDatabaseCommands {
-    /// List all databases in a fixed subscription
-    List {
-        /// Subscription ID
-        subscription_id: i32,
-    },
-    /// Get details of a specific fixed database
-    Get {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-    },
-    /// Create a new database in a fixed subscription
-    Create {
-        /// Subscription ID
-        subscription_id: i32,
-        /// JSON file with database configuration (use @filename or - for stdin)
-        file: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Update fixed database configuration
-    Update {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// JSON file with update configuration (use @filename or - for stdin)
-        file: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Delete a fixed database
-    Delete {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Skip confirmation prompt
-        #[arg(short, long)]
-        yes: bool,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Get backup status for fixed database
-    #[command(name = "backup-status")]
-    BackupStatus {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-    },
-    /// Trigger manual backup
-    Backup {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Get import status
-    #[command(name = "import-status")]
-    ImportStatus {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-    },
-    /// Import data into fixed database
-    Import {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// JSON file with import configuration (use @filename or - for stdin)
-        file: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Get slow query log
-    #[command(name = "slow-log")]
-    SlowLog {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Maximum number of entries to return
-        #[arg(long, default_value = "100")]
-        limit: i32,
-        /// Offset for pagination
-        #[arg(long, default_value = "0")]
-        offset: i32,
-    },
-    /// List tags for fixed database
-    #[command(name = "list-tags")]
-    ListTags {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-    },
-    /// Add a tag
-    #[command(name = "add-tag")]
-    AddTag {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Tag key
-        #[arg(long)]
-        key: String,
-        /// Tag value
-        #[arg(long)]
-        value: String,
-    },
-    /// Update all tags
-    #[command(name = "update-tags")]
-    UpdateTags {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// JSON file with tags (use @filename or - for stdin)
-        file: String,
-    },
-    /// Update specific tag
-    #[command(name = "update-tag")]
-    UpdateTag {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Tag key
-        #[arg(long)]
-        key: String,
-        /// Tag value
-        #[arg(long)]
-        value: String,
-    },
-    /// Delete a tag
-    #[command(name = "delete-tag")]
-    DeleteTag {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Tag key
-        #[arg(long)]
-        key: String,
-    },
-}
-
-/// Cloud Fixed Subscription Commands
-#[derive(Subcommand, Debug)]
-pub enum CloudFixedSubscriptionCommands {
-    /// List all available fixed subscription plans
-    #[command(name = "list-plans")]
-    ListPlans {
-        /// Filter by cloud provider (AWS, GCP, Azure)
-        #[arg(long)]
-        provider: Option<String>,
-    },
-    /// Get plans for a specific subscription
-    #[command(name = "get-plans")]
-    GetPlans {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-    },
-    /// Get details of a specific plan
-    #[command(name = "get-plan")]
-    GetPlan {
-        /// Plan ID
-        id: i32,
-    },
-    /// List all fixed subscriptions
-    List,
-    /// Get details of a fixed subscription
-    Get {
-        /// Subscription ID
-        id: i32,
-    },
-    /// Create a new fixed subscription
-    Create {
-        /// JSON file with subscription configuration (use @filename or - for stdin)
-        file: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Update fixed subscription
-    Update {
-        /// Subscription ID
-        id: i32,
-        /// JSON file with update configuration (use @filename or - for stdin)
-        file: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Delete a fixed subscription
-    Delete {
-        /// Subscription ID
-        id: i32,
-        /// Skip confirmation prompt
-        #[arg(short, long)]
-        yes: bool,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Get available Redis versions for fixed subscription
-    #[command(name = "redis-versions")]
-    RedisVersions {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: i32,
-    },
-}
-
-/// Cloud Provider Account Commands
-#[derive(Subcommand, Debug)]
-pub enum CloudProviderAccountCommands {
-    /// List all cloud provider accounts
-    List,
-    /// Get cloud provider account details
-    Get {
-        /// Cloud account ID
-        account_id: i32,
-    },
-    /// Create a new cloud provider account
-    Create {
-        /// JSON file containing the cloud account configuration
-        /// For GCP, this should be the service account JSON file
-        /// Use @filename to read from file
-        file: String,
-        /// Async operation arguments
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Update a cloud provider account
-    Update {
-        /// Cloud account ID
-        account_id: i32,
-        /// JSON file containing updated cloud account configuration
-        /// Use @filename to read from file
-        file: String,
-        /// Async operation arguments
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-    /// Delete a cloud provider account
-    Delete {
-        /// Cloud account ID
-        account_id: i32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-        /// Async operation arguments
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-}
-
-/// Cloud-specific commands (placeholder for now)
-#[derive(Subcommand, Debug)]
-pub enum CloudCommands {
-    /// Account operations
-    #[command(subcommand)]
-    Account(CloudAccountCommands),
-
-    /// Payment method operations
-    #[command(subcommand, name = "payment-method")]
-    PaymentMethod(CloudPaymentMethodCommands),
-
-    /// Subscription operations
-    #[command(subcommand)]
-    Subscription(CloudSubscriptionCommands),
-
-    /// Database operations
-    #[command(subcommand)]
-    Database(CloudDatabaseCommands),
-
-    /// User operations
-    #[command(subcommand)]
-    User(CloudUserCommands),
-
-    /// ACL (Access Control List) operations
-    #[command(subcommand)]
-    Acl(CloudAclCommands),
-    /// Cloud provider account operations
-    #[command(subcommand, name = "provider-account")]
-    ProviderAccount(CloudProviderAccountCommands),
-    /// Task operations
-    #[command(subcommand)]
-    Task(CloudTaskCommands),
-    /// Network connectivity operations (VPC, PSC, TGW)
-    #[command(subcommand)]
-    Connectivity(CloudConnectivityCommands),
-    /// Fixed database operations
-    #[command(subcommand, name = "fixed-database")]
-    FixedDatabase(CloudFixedDatabaseCommands),
-    /// Fixed subscription operations
-    #[command(subcommand, name = "fixed-subscription")]
-    FixedSubscription(CloudFixedSubscriptionCommands),
-    /// Workflow operations for multi-step tasks
-    #[command(subcommand)]
-    Workflow(CloudWorkflowCommands),
-}
-
-/// Enterprise-specific commands (placeholder for now)
-#[derive(Subcommand, Debug)]
-pub enum EnterpriseCommands {
-    /// Action (task) operations
-    #[command(subcommand)]
-    Action(crate::commands::enterprise::actions::ActionCommands),
-    /// Alert management operations
-    #[command(subcommand)]
-    Alerts(crate::commands::enterprise::alerts::AlertsCommands),
-    /// Database group operations
-    #[command(subcommand, name = "bdb-group")]
-    BdbGroup(crate::commands::enterprise::bdb_group::BdbGroupCommands),
-    /// Cluster operations
-    #[command(subcommand)]
-    Cluster(EnterpriseClusterCommands),
-    /// Cluster manager settings
-    #[command(subcommand, name = "cm-settings")]
-    CmSettings(crate::commands::enterprise::cm_settings::CmSettingsCommands),
-
-    /// Database operations
-    #[command(subcommand)]
-    Database(EnterpriseDatabaseCommands),
-
-    /// Debug info collection
-    #[command(subcommand)]
-    DebugInfo(crate::commands::enterprise::debuginfo::DebugInfoCommands),
-
-    /// Diagnostics operations
-    #[command(subcommand)]
-    Diagnostics(crate::commands::enterprise::diagnostics::DiagnosticsCommands),
-
-    /// Endpoint operations
-    #[command(subcommand)]
-    Endpoint(crate::commands::enterprise::endpoint::EndpointCommands),
-
-    /// Node operations
-    #[command(subcommand)]
-    Node(EnterpriseNodeCommands),
-
-    /// Proxy management
-    #[command(subcommand)]
-    Proxy(crate::commands::enterprise::proxy::ProxyCommands),
-
-    /// User operations
-    #[command(subcommand)]
-    User(EnterpriseUserCommands),
-
-    /// Role operations
-    #[command(subcommand)]
-    Role(EnterpriseRoleCommands),
-
-    /// ACL operations
-    #[command(subcommand)]
-    Acl(EnterpriseAclCommands),
-
-    /// LDAP integration
-    #[command(subcommand)]
-    Ldap(crate::commands::enterprise::ldap::LdapCommands),
-
-    /// LDAP mappings management
-    #[command(subcommand, name = "ldap-mappings")]
-    LdapMappings(crate::commands::enterprise::ldap::LdapMappingsCommands),
-
-    /// Authentication & sessions
-    #[command(subcommand)]
-    Auth(EnterpriseAuthCommands),
-    /// Bootstrap and initialization operations
-    #[command(subcommand)]
-    Bootstrap(crate::commands::enterprise::bootstrap::BootstrapCommands),
-
-    /// Active-Active database (CRDB) operations
-    #[command(subcommand)]
-    Crdb(EnterpriseCrdbCommands),
-    /// CRDB task operations
-    #[command(subcommand, name = "crdb-task")]
-    CrdbTask(crate::commands::enterprise::crdb_task::CrdbTaskCommands),
-
-    /// Job scheduler operations
-    #[command(subcommand, name = "job-scheduler")]
-    JobScheduler(crate::commands::enterprise::job_scheduler::JobSchedulerCommands),
-
-    /// JSON schema operations
-    #[command(subcommand)]
-    Jsonschema(crate::commands::enterprise::jsonschema::JsonSchemaCommands),
-
-    /// Log operations
-    #[command(subcommand)]
-    Logs(crate::commands::enterprise::logs::LogsCommands),
-    /// License management
-    #[command(subcommand)]
-    License(crate::commands::enterprise::license::LicenseCommands),
-
-    /// Migration operations
-    #[command(subcommand)]
-    Migration(crate::commands::enterprise::migration::MigrationCommands),
-
-    /// Module management operations
-    #[command(subcommand)]
-    Module(crate::commands::enterprise::module::ModuleCommands),
-
-    /// OCSP certificate validation
-    #[command(subcommand)]
-    Ocsp(crate::commands::enterprise::ocsp::OcspCommands),
-
-    /// Service management
-    #[command(subcommand)]
-    Services(crate::commands::enterprise::services::ServicesCommands),
-
-    /// Workflow operations for multi-step tasks
-    #[command(subcommand)]
-    Workflow(EnterpriseWorkflowCommands),
-
-    /// Local node operations
-    #[command(subcommand)]
-    Local(crate::commands::enterprise::local::LocalCommands),
-
-    /// Shard management operations
-    #[command(subcommand)]
-    Shard(crate::commands::enterprise::shard::ShardCommands),
-
-    /// Statistics and metrics operations
-    #[command(subcommand)]
-    Stats(EnterpriseStatsCommands),
-
-    /// Comprehensive cluster status (cluster, nodes, databases, shards)
-    Status {
-        /// Show only cluster information
-        #[arg(long)]
-        cluster: bool,
-
-        /// Show only nodes information
-        #[arg(long)]
-        nodes: bool,
-
-        /// Show only databases information
-        #[arg(long)]
-        databases: bool,
-
-        /// Show only shards information
-        #[arg(long)]
-        shards: bool,
-    },
-
-    /// Support package generation for troubleshooting
-    #[command(subcommand, name = "support-package")]
-    SupportPackage(crate::commands::enterprise::support_package::SupportPackageCommands),
-
-    /// DNS suffix management
-    #[command(subcommand)]
-    Suffix(crate::commands::enterprise::suffix::SuffixCommands),
-
-    /// Usage report operations
-    #[command(subcommand, name = "usage-report")]
-    UsageReport(crate::commands::enterprise::usage_report::UsageReportCommands),
-}
-
-/// Cloud workflow commands
-#[derive(Debug, Subcommand)]
-pub enum CloudWorkflowCommands {
-    /// List available workflows
-    List,
-    /// Complete subscription setup with optional database
-    #[command(name = "subscription-setup")]
-    SubscriptionSetup(crate::workflows::cloud::subscription_setup::SubscriptionSetupArgs),
-}
-
-/// Enterprise workflow commands
-#[derive(Debug, Subcommand)]
-pub enum EnterpriseWorkflowCommands {
-    /// List available workflows
-    List,
-    /// License management workflows
-    #[command(subcommand)]
-    License(crate::commands::enterprise::license_workflow::LicenseWorkflowCommands),
-
-    /// Initialize a Redis Enterprise cluster
-    #[command(name = "init-cluster")]
-    InitCluster {
-        /// Cluster name
-        #[arg(long, default_value = "redis-cluster")]
-        name: String,
-
-        /// Admin username
-        #[arg(long, default_value = "admin@redis.local")]
-        username: String,
-
-        /// Admin password (required)
-        #[arg(long, env = "REDIS_ENTERPRISE_INIT_PASSWORD")]
-        password: String,
-
-        /// Skip creating a default database after initialization
-        #[arg(long)]
-        skip_database: bool,
-
-        /// Name for the default database
-        #[arg(long, default_value = "default-db")]
-        database_name: String,
-
-        /// Memory size for the default database in GB
-        #[arg(long, default_value = "1")]
-        database_memory_gb: i64,
-
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-}
-
-// Placeholder command structures - will be expanded in later PRs
-
-#[derive(Subcommand, Debug)]
-pub enum CloudAccountCommands {
-    /// Get account information
-    Get,
-
-    /// Get payment methods configured for the account
-    GetPaymentMethods,
-
-    /// List supported regions
-    ListRegions {
-        /// Filter by cloud provider (aws, gcp, azure)
-        #[arg(long)]
-        provider: Option<String>,
-    },
-
-    /// List supported Redis modules
-    ListModules,
-
-    /// Get data persistence options
-    GetPersistenceOptions,
-
-    /// Get system logs
-    GetSystemLogs {
-        /// Maximum number of logs to return
-        #[arg(long, default_value = "100")]
-        limit: Option<u32>,
-
-        /// Offset for pagination
-        #[arg(long, default_value = "0")]
-        offset: Option<u32>,
-    },
-
-    /// Get session/audit logs
-    GetSessionLogs {
-        /// Maximum number of logs to return
-        #[arg(long, default_value = "100")]
-        limit: Option<u32>,
-
-        /// Offset for pagination
-        #[arg(long, default_value = "0")]
-        offset: Option<u32>,
-    },
-
-    /// Get search module scaling factors
-    GetSearchScaling,
-}
-
-#[derive(Subcommand, Debug)]
-pub enum CloudPaymentMethodCommands {
-    /// List payment methods configured for the account
-    List,
-}
-
-#[derive(Subcommand, Debug)]
-pub enum CloudSubscriptionCommands {
-    /// List all subscriptions
-    List,
-
-    /// Get detailed subscription information
-    Get {
-        /// Subscription ID
-        id: u32,
-    },
-
-    /// Create a new subscription
-    #[command(after_help = "EXAMPLES:
-    # Simple subscription - just name, provider, and region via --data
-    redisctl cloud subscription create --name prod-subscription \\
-      --data '{\"cloudProviders\":[{\"regions\":[{\"region\":\"us-east-1\"}]}],\"databases\":[{\"name\":\"db1\",\"memoryLimitInGb\":1}]}'
-
-    # With payment method
-    redisctl cloud subscription create --name dev-subscription \\
-      --payment-method marketplace \\
-      --data '{\"cloudProviders\":[{\"regions\":[{\"region\":\"us-west-2\"}]}],\"databases\":[{\"name\":\"db1\",\"memoryLimitInGb\":1}]}'
-
-    # With auto-tiering (RAM+Flash)
-    redisctl cloud subscription create --name large-subscription \\
-      --memory-storage ram-and-flash \\
-      --data '{\"cloudProviders\":[{\"provider\":\"AWS\",\"regions\":[{\"region\":\"eu-west-1\"}]}],\"databases\":[{\"name\":\"db1\",\"memoryLimitInGb\":10}]}'
-
-    # Complete configuration from file
-    redisctl cloud subscription create --data @subscription.json
-
-    # Dry run to preview deployment
-    redisctl cloud subscription create --dry-run --data @subscription.json
-
-NOTE: Subscription creation requires complex nested structures for cloud providers,
-      regions, and databases. Use --data for the required cloudProviders and databases
-      arrays. First-class parameters (--name, --payment-method, etc.) override values
-      in --data when both are provided.")]
-    Create {
-        /// Subscription name
-        #[arg(long)]
-        name: Option<String>,
-
-        /// Dry run - create deployment plan without provisioning resources
-        #[arg(long)]
-        dry_run: bool,
-
-        /// Deployment type: single-region or active-active
-        #[arg(long, value_parser = ["single-region", "active-active"])]
-        deployment_type: Option<String>,
-
-        /// Payment method: credit-card or marketplace
-        #[arg(long, value_parser = ["credit-card", "marketplace"], default_value = "credit-card")]
-        payment_method: String,
-
-        /// Payment method ID (required if payment-method is credit-card)
-        #[arg(long)]
-        payment_method_id: Option<i32>,
-
-        /// Memory storage: ram or ram-and-flash (Auto Tiering)
-        #[arg(long, value_parser = ["ram", "ram-and-flash"], default_value = "ram")]
-        memory_storage: String,
-
-        /// Persistent storage encryption: cloud-provider-managed-key or customer-managed-key
-        #[arg(long, value_parser = ["cloud-provider-managed-key", "customer-managed-key"], default_value = "cloud-provider-managed-key")]
-        persistent_storage_encryption: String,
-
-        /// Advanced: Full subscription configuration as JSON string or @file.json
-        /// REQUIRED: Must include cloudProviders array with regions and databases array
-        #[arg(long)]
-        data: Option<String>,
-
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Update subscription configuration
-    Update {
-        /// Subscription ID
-        id: u32,
-        /// Update configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Delete a subscription
-    Delete {
-        /// Subscription ID
-        id: u32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Get available Redis versions
-    RedisVersions {
-        /// Filter by subscription ID (optional)
-        #[arg(long)]
-        subscription: Option<u32>,
-    },
-
-    /// Get subscription pricing information
-    GetPricing {
-        /// Subscription ID
-        id: u32,
-    },
-
-    /// Get CIDR allowlist
-    GetCidrAllowlist {
-        /// Subscription ID
-        id: u32,
-    },
-
-    /// Update CIDR allowlist
-    UpdateCidrAllowlist {
-        /// Subscription ID
-        id: u32,
-        /// CIDR blocks as JSON array or @file.json
-        #[arg(long)]
-        cidrs: String,
-    },
-
-    /// Get maintenance windows
-    GetMaintenanceWindows {
-        /// Subscription ID
-        id: u32,
-    },
-
-    /// Update maintenance windows
-    UpdateMaintenanceWindows {
-        /// Subscription ID
-        id: u32,
-        /// Maintenance windows configuration as JSON or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// List Active-Active regions
-    ListAaRegions {
-        /// Subscription ID
-        id: u32,
-    },
-
-    /// Add region to Active-Active subscription
-    AddAaRegion {
-        /// Subscription ID
-        id: u32,
-        /// Region configuration as JSON or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Delete regions from Active-Active subscription
-    DeleteAaRegions {
-        /// Subscription ID
-        id: u32,
-        /// Regions to delete as JSON array or @file.json
-        #[arg(long)]
-        regions: String,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum CloudDatabaseCommands {
-    /// List all databases across subscriptions
-    List {
-        /// Filter by subscription ID
-        #[arg(long)]
-        subscription: Option<u32>,
-    },
-
-    /// Get detailed database information
-    Get {
-        /// Database ID (format: subscription_id:database_id for fixed, or just database_id for flexible)
-        id: String,
-    },
-
-    /// Create a new database
-    #[command(after_help = "EXAMPLES:
-    # Simple database - just name and size
-    redisctl cloud database create --subscription 123 --name mydb --memory 1
-
-    # Production database with high availability
-    redisctl cloud database create \\
-      --subscription 123 \\
-      --name prod-cache \\
-      --memory 10 \\
-      --replication \\
-      --data-persistence aof-every-1-second
-
-    # Advanced: Mix flags with JSON for rare options
-    redisctl cloud database create \\
-      --subscription 123 \\
-      --name mydb \\
-      --memory 5 \\
-      --data '{\"modules\": [{\"name\": \"RedisJSON\"}]}'
-")]
-    Create {
-        /// Subscription ID
-        #[arg(long)]
-        subscription: u32,
-
-        /// Database name (required unless using --data)
-        /// Limited to 40 characters: letters, digits, hyphens
-        /// Must start with letter, end with letter or digit
-        #[arg(long)]
-        name: Option<String>,
-
-        /// Memory limit in GB (e.g., 1, 5, 10, 50)
-        /// Alternative to --dataset-size
-        #[arg(long, conflicts_with = "dataset_size")]
-        memory: Option<f64>,
-
-        /// Dataset size in GB (alternative to --memory)
-        /// If replication enabled, total memory will be 2x this value
-        #[arg(long, conflicts_with = "memory")]
-        dataset_size: Option<f64>,
-
-        /// Database protocol
-        #[arg(long, value_parser = ["redis", "memcached"], default_value = "redis")]
-        protocol: String,
-
-        /// Enable replication for high availability
-        #[arg(long)]
-        replication: bool,
-
-        /// Data persistence policy
-        /// Options: none, aof-every-1-second, aof-every-write, snapshot-every-1-hour,
-        ///          snapshot-every-6-hours, snapshot-every-12-hours
-        #[arg(long)]
-        data_persistence: Option<String>,
-
-        /// Data eviction policy when memory limit reached
-        /// Options: volatile-lru, volatile-ttl, volatile-random, allkeys-lru,
-        ///          allkeys-lfu, allkeys-random, noeviction, volatile-lfu
-        #[arg(long, default_value = "volatile-lru")]
-        eviction_policy: String,
-
-        /// Redis version (e.g., "7.2", "7.0", "6.2")
-        #[arg(long)]
-        redis_version: Option<String>,
-
-        /// Enable OSS Cluster API support
-        #[arg(long)]
-        oss_cluster: bool,
-
-        /// TCP port (10000-19999, auto-assigned if not specified)
-        #[arg(long)]
-        port: Option<i32>,
-
-        /// Advanced: Full database configuration as JSON string or @file.json
-        /// CLI flags take precedence over values in JSON
-        #[arg(long)]
-        data: Option<String>,
-
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Update database configuration
-    Update {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Update configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Delete a database
-    Delete {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Get database backup status
-    BackupStatus {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-    },
-
-    /// Trigger manual database backup
-    Backup {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Get database import status
-    ImportStatus {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-    },
-
-    /// Import data into database
-    Import {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Import configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Get database certificate
-    GetCertificate {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-    },
-
-    /// Get slow query log
-    SlowLog {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Maximum number of entries to return
-        #[arg(long, default_value = "100")]
-        limit: u32,
-        /// Offset for pagination
-        #[arg(long, default_value = "0")]
-        offset: u32,
-    },
-
-    /// List database tags
-    ListTags {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-    },
-
-    /// Add a tag to database
-    AddTag {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Tag key
-        #[arg(long)]
-        key: String,
-        /// Tag value
-        #[arg(long)]
-        value: String,
-    },
-
-    /// Update database tags
-    UpdateTags {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Tags as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Delete a tag from database
-    DeleteTag {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Tag key to delete
-        #[arg(long)]
-        key: String,
-    },
-
-    /// Flush Active-Active database
-    FlushCrdb {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-
-    /// Get Redis version upgrade status
-    UpgradeStatus {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-    },
-
-    /// Upgrade Redis version
-    UpgradeRedis {
-        /// Database ID (format: subscription_id:database_id)
-        id: String,
-        /// Target Redis version
-        #[arg(long)]
-        version: String,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum CloudUserCommands {
-    /// List all users
-    List,
-
-    /// Get detailed user information
-    Get {
-        /// User ID
-        id: u32,
-    },
-
-    /// Update user information
-    Update {
-        /// User ID
-        id: u32,
-        /// New name for the user
-        #[arg(long)]
-        name: Option<String>,
-        /// New role for the user (owner, manager, viewer, billing_admin)
-        #[arg(long)]
-        role: Option<String>,
-        /// Enable/disable email alerts
-        #[arg(long)]
-        alerts_email: Option<bool>,
-        /// Enable/disable SMS alerts
-        #[arg(long)]
-        alerts_sms: Option<bool>,
-    },
-
-    /// Delete a user
-    Delete {
-        /// User ID
-        id: u32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-        /// Async operation arguments
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum CloudAclCommands {
-    // Redis ACL Rules
-    /// List all Redis ACL rules
-    #[command(name = "list-redis-rules")]
-    ListRedisRules,
-
-    /// Create a new Redis ACL rule
-    #[command(name = "create-redis-rule")]
-    CreateRedisRule {
-        /// Rule name
-        #[arg(long)]
-        name: String,
-        /// Redis ACL rule (e.g., "+@read")
-        #[arg(long)]
-        rule: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Update an existing Redis ACL rule
-    #[command(name = "update-redis-rule")]
-    UpdateRedisRule {
-        /// Rule ID
-        id: i32,
-        /// New rule name
-        #[arg(long)]
-        name: Option<String>,
-        /// New Redis ACL rule
-        #[arg(long)]
-        rule: Option<String>,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Delete a Redis ACL rule
-    #[command(name = "delete-redis-rule")]
-    DeleteRedisRule {
-        /// Rule ID
-        id: i32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    // ACL Roles
-    /// List all ACL roles
-    #[command(name = "list-roles")]
-    ListRoles,
-
-    /// Create a new ACL role
-    #[command(name = "create-role")]
-    CreateRole {
-        /// Role name
-        #[arg(long)]
-        name: String,
-        /// Redis rules (JSON array or single rule ID)
-        #[arg(long, value_name = "JSON|ID")]
-        redis_rules: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Update an existing ACL role
-    #[command(name = "update-role")]
-    UpdateRole {
-        /// Role ID
-        id: i32,
-        /// New role name
-        #[arg(long)]
-        name: Option<String>,
-        /// New Redis rules (JSON array or single rule ID)
-        #[arg(long, value_name = "JSON|ID")]
-        redis_rules: Option<String>,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Delete an ACL role
-    #[command(name = "delete-role")]
-    DeleteRole {
-        /// Role ID
-        id: i32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    // ACL Users
-    /// List all ACL users
-    #[command(name = "list-acl-users")]
-    ListAclUsers,
-
-    /// Get ACL user details
-    #[command(name = "get-acl-user")]
-    GetAclUser {
-        /// ACL user ID
-        id: i32,
-    },
-
-    /// Create a new ACL user
-    #[command(name = "create-acl-user")]
-    CreateAclUser {
-        /// Username
-        #[arg(long)]
-        name: String,
-        /// Role name
-        #[arg(long)]
-        role: String,
-        /// Password
-        #[arg(long)]
-        password: String,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Update an ACL user
-    #[command(name = "update-acl-user")]
-    UpdateAclUser {
-        /// ACL user ID
-        id: i32,
-        /// New username
-        #[arg(long)]
-        name: Option<String>,
-        /// New role name
-        #[arg(long)]
-        role: Option<String>,
-        /// New password
-        #[arg(long)]
-        password: Option<String>,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-
-    /// Delete an ACL user
-    #[command(name = "delete-acl-user")]
-    DeleteAclUser {
-        /// ACL user ID
-        id: i32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-        /// Async operation options
-        #[command(flatten)]
-        async_ops: crate::commands::cloud::async_utils::AsyncOperationArgs,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum EnterpriseClusterCommands {
-    /// Get cluster configuration
-    Get,
-
-    /// Update cluster configuration
-    Update {
-        /// Cluster configuration data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Get cluster policies
-    #[command(name = "get-policy")]
-    GetPolicy,
-
-    /// Update cluster policies
-    #[command(name = "update-policy")]
-    UpdatePolicy {
-        /// Policy data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Get license information
-    #[command(name = "get-license")]
-    GetLicense,
-
-    /// Update license
-    #[command(name = "update-license")]
-    UpdateLicense {
-        /// License key file or content
-        #[arg(long, value_name = "FILE|KEY")]
-        license: String,
-    },
-
-    /// Bootstrap new cluster
-    Bootstrap {
-        /// Bootstrap configuration (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Join node to cluster
-    Join {
-        /// Join configuration (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Recover cluster
-    Recover {
-        /// Recovery configuration (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Reset cluster (dangerous!)
-    Reset {
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-
-    /// Get cluster statistics
-    Stats,
-
-    /// Get cluster metrics
-    Metrics {
-        /// Time interval (e.g., "1h", "5m")
-        #[arg(long)]
-        interval: Option<String>,
-    },
-
-    /// Get active alerts
-    Alerts,
-
-    /// Get cluster events
-    Events {
-        /// Maximum number of events to return
-        #[arg(long, default_value = "100")]
-        limit: Option<u32>,
-    },
-
-    /// Get audit log
-    #[command(name = "audit-log")]
-    AuditLog {
-        /// From date (e.g., "2024-01-01")
-        #[arg(long)]
-        from: Option<String>,
-    },
-
-    /// Enable maintenance mode
-    #[command(name = "maintenance-mode-enable")]
-    MaintenanceModeEnable,
-
-    /// Disable maintenance mode
-    #[command(name = "maintenance-mode-disable")]
-    MaintenanceModeDisable,
-
-    /// Collect debug information
-    #[command(name = "debug-info")]
-    DebugInfo,
-
-    /// Check cluster health status
-    #[command(name = "check-status")]
-    CheckStatus,
-
-    /// Get cluster certificates
-    #[command(name = "get-certificates")]
-    GetCertificates,
-
-    /// Update certificates
-    #[command(name = "update-certificates")]
-    UpdateCertificates {
-        /// Certificate data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Rotate certificates
-    #[command(name = "rotate-certificates")]
-    RotateCertificates,
-
-    /// Get OCSP configuration
-    #[command(name = "get-ocsp")]
-    GetOcsp,
-
-    /// Update OCSP configuration
-    #[command(name = "update-ocsp")]
-    UpdateOcsp {
-        /// OCSP configuration data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum EnterpriseDatabaseCommands {
-    /// List all databases
-    List,
-
-    /// Get database details
-    Get {
-        /// Database ID
-        id: u32,
-    },
-
-    /// Create a new database
-    #[command(after_help = "EXAMPLES:
-    # Simple database - just name and size
-    redisctl enterprise database create --name mydb --memory 1073741824
-
-    # With replication for high availability
-    redisctl enterprise database create --name prod-db --memory 2147483648 --replication
-
-    # With persistence and eviction policy
-    redisctl enterprise database create --name cache-db --memory 536870912 \\
-      --persistence aof --eviction-policy volatile-lru
-
-    # With sharding for horizontal scaling
-    redisctl enterprise database create --name large-db --memory 10737418240 \\
-      --sharding --shards-count 4
-
-    # With specific port
-    redisctl enterprise database create --name service-db --memory 1073741824 --port 12000
-
-    # Complete configuration from file
-    redisctl enterprise database create --data @database.json
-
-    # Dry run to preview without creating
-    redisctl enterprise database create --name test-db --memory 1073741824 --dry-run
-
-NOTE: Memory size is in bytes. Common values:
-      - 1 GB = 1073741824 bytes
-      - 2 GB = 2147483648 bytes
-      - 5 GB = 5368709120 bytes
-      First-class parameters override values in --data when both are provided.")]
-    Create {
-        /// Database name (required unless using --data)
-        #[arg(long)]
-        name: Option<String>,
-
-        /// Memory size in bytes (e.g., 1073741824 for 1GB)
-        #[arg(long)]
-        memory: Option<u64>,
-
-        /// TCP port (10000-19999, auto-assigned if not specified)
-        #[arg(long)]
-        port: Option<u16>,
-
-        /// Enable replication for high availability
-        #[arg(long)]
-        replication: bool,
-
-        /// Data persistence: aof, snapshot, or aof-and-snapshot
-        #[arg(long)]
-        persistence: Option<String>,
-
-        /// Data eviction policy when memory limit reached
-        #[arg(long)]
-        eviction_policy: Option<String>,
-
-        /// Enable sharding for horizontal scaling
-        #[arg(long)]
-        sharding: bool,
-
-        /// Number of shards (requires --sharding)
-        #[arg(long)]
-        shards_count: Option<u32>,
-
-        /// Proxy policy: single, all-master-shards, or all-nodes
-        #[arg(long)]
-        proxy_policy: Option<String>,
-
-        /// Enable CRDB (Active-Active)
-        #[arg(long)]
-        crdb: bool,
-
-        /// Redis password for authentication
-        #[arg(long)]
-        redis_password: Option<String>,
-
-        /// Advanced: Full database configuration as JSON string or @file.json
-        #[arg(long)]
-        data: Option<String>,
-
-        /// Perform a dry run without creating the database
-        #[arg(long)]
-        dry_run: bool,
-    },
-
-    /// Update database configuration
-    Update {
-        /// Database ID
-        id: u32,
-        /// Update configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Delete a database
-    Delete {
-        /// Database ID
-        id: u32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-
-    /// Export database
-    Export {
-        /// Database ID
-        id: u32,
-        /// Export configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Import to database
-    Import {
-        /// Database ID
-        id: u32,
-        /// Import configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Trigger database backup
-    Backup {
-        /// Database ID
-        id: u32,
-    },
-
-    /// Restore database
-    Restore {
-        /// Database ID
-        id: u32,
-        /// Restore configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Flush database data
-    Flush {
-        /// Database ID
-        id: u32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-
-    /// Get database shards info
-    GetShards {
-        /// Database ID
-        id: u32,
-    },
-
-    /// Update sharding configuration
-    UpdateShards {
-        /// Database ID
-        id: u32,
-        /// Shards configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Get enabled modules
-    GetModules {
-        /// Database ID
-        id: u32,
-    },
-
-    /// Update modules configuration
-    UpdateModules {
-        /// Database ID
-        id: u32,
-        /// Modules configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Upgrade database Redis version
-    Upgrade {
-        /// Database ID
-        id: u32,
-        /// Target Redis version (defaults to latest)
-        #[arg(long)]
-        version: Option<String>,
-        /// Preserve master/replica roles (requires extra failover)
-        #[arg(long)]
-        preserve_roles: bool,
-        /// Restart shards even if no version change
-        #[arg(long)]
-        force_restart: bool,
-        /// Allow data loss in non-replicated, non-persistent databases
-        #[arg(long)]
-        may_discard_data: bool,
-        /// Force data discard even if replicated/persistent
-        #[arg(long)]
-        force_discard: bool,
-        /// Keep current CRDT protocol version
-        #[arg(long)]
-        keep_crdt_protocol_version: bool,
-        /// Maximum parallel shard upgrades
-        #[arg(long)]
-        parallel_shards_upgrade: Option<u32>,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-
-    /// Get ACL configuration
-    GetAcl {
-        /// Database ID
-        id: u32,
-    },
-
-    /// Update ACL configuration
-    UpdateAcl {
-        /// Database ID
-        id: u32,
-        /// ACL configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Get database statistics
-    Stats {
-        /// Database ID
-        id: u32,
-    },
-
-    /// Get database metrics
-    Metrics {
-        /// Database ID
-        id: u32,
-        /// Time interval (e.g., "1h", "24h")
-        #[arg(long)]
-        interval: Option<String>,
-    },
-
-    /// Get slow query log
-    Slowlog {
-        /// Database ID
-        id: u32,
-        /// Limit number of entries
-        #[arg(long)]
-        limit: Option<u32>,
-    },
-
-    /// Get connected clients
-    ClientList {
-        /// Database ID
-        id: u32,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum EnterpriseNodeCommands {
-    /// List all nodes in cluster
-    List,
-
-    /// Get node details
-    Get {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Add node to cluster
-    Add {
-        /// Node configuration (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Remove node from cluster
-    Remove {
-        /// Node ID
-        id: u32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-
-    /// Update node configuration
-    Update {
-        /// Node ID
-        id: u32,
-        /// Update data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Get node status
-    Status {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Get node statistics
-    Stats {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Get node metrics
-    Metrics {
-        /// Node ID
-        id: u32,
-        /// Time interval (e.g., "1h", "5m")
-        #[arg(long)]
-        interval: Option<String>,
-    },
-
-    /// Run health check on node
-    Check {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Get node-specific alerts
-    Alerts {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Put node in maintenance mode
-    #[command(name = "maintenance-enable")]
-    MaintenanceEnable {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Remove node from maintenance mode
-    #[command(name = "maintenance-disable")]
-    MaintenanceDisable {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Rebalance shards on node
-    Rebalance {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Drain node before removal
-    Drain {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Restart node services
-    Restart {
-        /// Node ID
-        id: u32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-
-    /// Get node configuration
-    #[command(name = "get-config")]
-    GetConfig {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Update node configuration
-    #[command(name = "update-config")]
-    UpdateConfig {
-        /// Node ID
-        id: u32,
-        /// Configuration data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Get rack awareness configuration
-    #[command(name = "get-rack")]
-    GetRack {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Set rack ID
-    #[command(name = "set-rack")]
-    SetRack {
-        /// Node ID
-        id: u32,
-        /// Rack identifier
-        #[arg(long)]
-        rack: String,
-    },
-
-    /// Get node role
-    #[command(name = "get-role")]
-    GetRole {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Get resource utilization
-    Resources {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Get memory usage details
-    Memory {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Get CPU usage details
-    Cpu {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Get storage usage details
-    Storage {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Get network statistics
-    Network {
-        /// Node ID
-        id: u32,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum EnterpriseUserCommands {
-    /// List all users
-    List,
-
-    /// Get user details
-    Get {
-        /// User ID
-        id: u32,
-    },
-
-    /// Create new user
-    Create {
-        /// User data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Update user
-    Update {
-        /// User ID
-        id: u32,
-        /// Update data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Delete user
-    Delete {
-        /// User ID
-        id: u32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-
-    /// Reset user password
-    #[command(name = "reset-password")]
-    ResetPassword {
-        /// User ID
-        id: u32,
-        /// New password (will prompt if not provided)
-        #[arg(long)]
-        password: Option<String>,
-    },
-
-    /// Get user's roles
-    #[command(name = "get-roles")]
-    GetRoles {
-        /// User ID
-        #[arg(name = "user-id")]
-        user_id: u32,
-    },
-
-    /// Assign role to user
-    #[command(name = "assign-role")]
-    AssignRole {
-        /// User ID
-        #[arg(name = "user-id")]
-        user_id: u32,
-        /// Role ID to assign
-        #[arg(long)]
-        role: u32,
-    },
-
-    /// Remove role from user
-    #[command(name = "remove-role")]
-    RemoveRole {
-        /// User ID
-        #[arg(name = "user-id")]
-        user_id: u32,
-        /// Role ID to remove
-        #[arg(long)]
-        role: u32,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum EnterpriseRoleCommands {
-    /// List all roles
-    List,
-
-    /// Get role details
-    Get {
-        /// Role ID
-        id: u32,
-    },
-
-    /// Create custom role
-    Create {
-        /// Role data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Update role
-    Update {
-        /// Role ID
-        id: u32,
-        /// Update data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Delete custom role
-    Delete {
-        /// Role ID
-        id: u32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-
-    /// Get role permissions
-    #[command(name = "get-permissions")]
-    GetPermissions {
-        /// Role ID
-        id: u32,
-    },
-
-    /// Get users with specific role
-    #[command(name = "get-users")]
-    GetUsers {
-        /// Role ID
-        #[arg(name = "role-id")]
-        role_id: u32,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum EnterpriseAclCommands {
-    /// List all ACLs
-    List,
-
-    /// Get ACL details
-    Get {
-        /// ACL ID
-        id: u32,
-    },
-
-    /// Create ACL
-    Create {
-        /// ACL data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Update ACL
-    Update {
-        /// ACL ID
-        id: u32,
-        /// Update data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Delete ACL
-    Delete {
-        /// ACL ID
-        id: u32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-
-    /// Test ACL permissions
-    Test {
-        /// User ID
-        #[arg(long)]
-        user: u32,
-        /// Redis command to test
-        #[arg(long)]
-        command: String,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum EnterpriseLdapCommands {
-    /// Get LDAP configuration
-    #[command(name = "get-config")]
-    GetConfig,
-
-    /// Update LDAP configuration
-    #[command(name = "update-config")]
-    UpdateConfig {
-        /// LDAP config data (JSON file or inline)
-        #[arg(long, value_name = "FILE|JSON")]
-        data: String,
-    },
-
-    /// Test LDAP connection
-    #[command(name = "test-connection")]
-    TestConnection,
-
-    /// Sync users from LDAP
-    Sync,
-
-    /// Get LDAP role mappings
-    #[command(name = "get-mappings")]
-    GetMappings,
-}
-
-#[derive(Subcommand, Debug)]
-pub enum EnterpriseAuthCommands {
-    /// Test authentication
-    Test {
-        /// Username/email to test
-        #[arg(long)]
-        user: String,
-    },
-
-    /// List active sessions
-    #[command(name = "session-list")]
-    SessionList,
-
-    /// Revoke session
-    #[command(name = "session-revoke")]
-    SessionRevoke {
-        /// Session ID
-        #[arg(name = "session-id")]
-        session_id: String,
-    },
-
-    /// Revoke all user sessions
-    #[command(name = "session-revoke-all")]
-    SessionRevokeAll {
-        /// User ID
-        #[arg(long)]
-        user: u32,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum EnterpriseCrdbCommands {
-    // CRDB Lifecycle Management
-    /// List all Active-Active databases
-    List,
-
-    /// Get CRDB details
-    Get {
-        /// CRDB ID
-        id: u32,
-    },
-
-    /// Create Active-Active database
-    Create {
-        /// CRDB configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Update CRDB configuration
-    Update {
-        /// CRDB ID
-        id: u32,
-        /// Update configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Delete CRDB
-    Delete {
-        /// CRDB ID
-        id: u32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-
-    // Participating Clusters Management
-    /// Get participating clusters
-    #[command(name = "get-clusters")]
-    GetClusters {
-        /// CRDB ID
-        id: u32,
-    },
-
-    /// Add cluster to CRDB
-    #[command(name = "add-cluster")]
-    AddCluster {
-        /// CRDB ID
-        id: u32,
-        /// Cluster configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Remove cluster from CRDB
-    #[command(name = "remove-cluster")]
-    RemoveCluster {
-        /// CRDB ID
-        id: u32,
-        /// Cluster ID to remove
-        #[arg(long)]
-        cluster: u32,
-    },
-
-    /// Update cluster configuration in CRDB
-    #[command(name = "update-cluster")]
-    UpdateCluster {
-        /// CRDB ID
-        id: u32,
-        /// Cluster ID to update
-        #[arg(long)]
-        cluster: u32,
-        /// Update configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    // Instance Management
-    /// Get all CRDB instances
-    #[command(name = "get-instances")]
-    GetInstances {
-        /// CRDB ID
-        id: u32,
-    },
-
-    /// Get specific CRDB instance
-    #[command(name = "get-instance")]
-    GetInstance {
-        /// CRDB ID
-        #[arg(name = "crdb-id")]
-        crdb_id: u32,
-        /// Instance ID
-        #[arg(long)]
-        instance: u32,
-    },
-
-    /// Update CRDB instance
-    #[command(name = "update-instance")]
-    UpdateInstance {
-        /// CRDB ID
-        #[arg(name = "crdb-id")]
-        crdb_id: u32,
-        /// Instance ID
-        #[arg(long)]
-        instance: u32,
-        /// Update configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Flush CRDB instance data
-    #[command(name = "flush-instance")]
-    FlushInstance {
-        /// CRDB ID
-        #[arg(name = "crdb-id")]
-        crdb_id: u32,
-        /// Instance ID
-        #[arg(long)]
-        instance: u32,
-        /// Skip confirmation prompt
-        #[arg(long)]
-        force: bool,
-    },
-
-    // Replication & Sync
-    /// Get replication status
-    #[command(name = "get-replication-status")]
-    GetReplicationStatus {
-        /// CRDB ID
-        id: u32,
-    },
-
-    /// Get replication lag metrics
-    #[command(name = "get-lag")]
-    GetLag {
-        /// CRDB ID
-        id: u32,
-    },
-
-    /// Force synchronization
-    #[command(name = "force-sync")]
-    ForceSync {
-        /// CRDB ID
-        id: u32,
-        /// Source cluster ID
-        #[arg(long)]
-        source: u32,
-    },
-
-    /// Pause replication
-    #[command(name = "pause-replication")]
-    PauseReplication {
-        /// CRDB ID
-        id: u32,
-    },
-
-    /// Resume replication
-    #[command(name = "resume-replication")]
-    ResumeReplication {
-        /// CRDB ID
-        id: u32,
-    },
-
-    // Conflict Resolution
-    /// Get conflict history
-    #[command(name = "get-conflicts")]
-    GetConflicts {
-        /// CRDB ID
-        id: u32,
-        /// Maximum number of conflicts to return
-        #[arg(long)]
-        limit: Option<u32>,
-    },
-
-    /// Get conflict resolution policy
-    #[command(name = "get-conflict-policy")]
-    GetConflictPolicy {
-        /// CRDB ID
-        id: u32,
-    },
-
-    /// Update conflict resolution policy
-    #[command(name = "update-conflict-policy")]
-    UpdateConflictPolicy {
-        /// CRDB ID
-        id: u32,
-        /// Policy configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Manually resolve conflict
-    #[command(name = "resolve-conflict")]
-    ResolveConflict {
-        /// CRDB ID
-        id: u32,
-        /// Conflict ID
-        #[arg(long)]
-        conflict: String,
-        /// Resolution method
-        #[arg(long)]
-        resolution: String,
-    },
-
-    // Tasks & Jobs
-    /// Get CRDB tasks
-    #[command(name = "get-tasks")]
-    GetTasks {
-        /// CRDB ID
-        id: u32,
-    },
-
-    /// Get specific task details
-    #[command(name = "get-task")]
-    GetTask {
-        /// CRDB ID
-        #[arg(name = "crdb-id")]
-        crdb_id: u32,
-        /// Task ID
-        #[arg(long)]
-        task: String,
-    },
-
-    /// Retry failed task
-    #[command(name = "retry-task")]
-    RetryTask {
-        /// CRDB ID
-        #[arg(name = "crdb-id")]
-        crdb_id: u32,
-        /// Task ID
-        #[arg(long)]
-        task: String,
-    },
-
-    /// Cancel running task
-    #[command(name = "cancel-task")]
-    CancelTask {
-        /// CRDB ID
-        #[arg(name = "crdb-id")]
-        crdb_id: u32,
-        /// Task ID
-        #[arg(long)]
-        task: String,
-    },
-
-    // Monitoring & Metrics
-    /// Get CRDB statistics
-    Stats {
-        /// CRDB ID
-        id: u32,
-    },
-
-    /// Get CRDB metrics
-    Metrics {
-        /// CRDB ID
-        id: u32,
-        /// Time interval (e.g., "1h", "24h")
-        #[arg(long)]
-        interval: Option<String>,
-    },
-
-    /// Get connection details per instance
-    #[command(name = "get-connections")]
-    GetConnections {
-        /// CRDB ID
-        id: u32,
-    },
-
-    /// Get throughput metrics
-    #[command(name = "get-throughput")]
-    GetThroughput {
-        /// CRDB ID
-        id: u32,
-    },
-
-    /// Run health check
-    #[command(name = "health-check")]
-    HealthCheck {
-        /// CRDB ID
-        id: u32,
-    },
-
-    // Backup & Recovery
-    /// Create CRDB backup
-    Backup {
-        /// CRDB ID
-        id: u32,
-        /// Backup configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// Restore CRDB
-    Restore {
-        /// CRDB ID
-        id: u32,
-        /// Restore configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-
-    /// List available backups
-    #[command(name = "get-backups")]
-    GetBackups {
-        /// CRDB ID
-        id: u32,
-    },
-
-    /// Export CRDB data
-    Export {
-        /// CRDB ID
-        id: u32,
-        /// Export configuration as JSON string or @file.json
-        #[arg(long)]
-        data: String,
-    },
-}
-
-#[derive(Subcommand, Debug, Clone)]
-pub enum EnterpriseStatsCommands {
-    /// Get database statistics
-    Database {
-        /// Database ID
-        id: u32,
-    },
-
-    /// Get database shard statistics
-    DatabaseShards {
-        /// Database ID
-        id: u32,
-    },
-
-    /// Get database metrics over time
-    DatabaseMetrics {
-        /// Database ID
-        id: u32,
-        /// Time interval (1m, 5m, 1h, 1d)
-        #[arg(long, default_value = "1h")]
-        interval: String,
-    },
-
-    /// Get node statistics
-    Node {
-        /// Node ID
-        id: u32,
-    },
-
-    /// Get node metrics over time
-    NodeMetrics {
-        /// Node ID
-        id: u32,
-        /// Time interval (1m, 5m, 1h, 1d)
-        #[arg(long, default_value = "1h")]
-        interval: String,
-    },
-
-    /// Get cluster-wide statistics
-    Cluster,
-
-    /// Get cluster metrics over time
-    ClusterMetrics {
-        /// Time interval (1m, 5m, 1h, 1d)
-        #[arg(long, default_value = "1h")]
-        interval: String,
-    },
-
-    /// Get listener statistics
-    Listener,
-
-    /// Export statistics in various formats
-    Export {
-        /// Export format (json, prometheus, csv)
-        #[arg(long, default_value = "json")]
-        format: String,
-        /// Time interval for time-series data (1m, 5m, 1h, 1d)
-        #[arg(long)]
-        interval: Option<String>,
-    },
-}
